rr_graph: Split utils into own files
