+++ utils/lib/rr_graph/graph.py
@@ -789,6 +789,7 @@ class Block(MostlyReadOnly):
         else:
             raise KeyError("ptc %d not found" % ptc)
 
+
 class BlockTypeEdge(enum.Enum):
     TOP = "TOP"
     LEFT = "LEFT"
@@ -1792,98 +1793,12 @@ def test_index_node_objects():
     for track, node in track2node.items():
         assert track.idx == node_ptc(node)
 
-def print_block_types(rr_graph):
-    '''Sequentially list block types'''
-    bg = rr_graph.block_grid
-
-    for type_id, bt in bg.block_types.items():
-        print("{:4}  ".format(type_id), "{:40s}".format(bt.to_string()), bt.to_string(extra=True))
-
-def print_grid(rr_graph):
-    '''ASCII diagram displaying XY layout'''
-    bg = rr_graph.block_grid
-    grid = bg.size()
-
-    #print('Grid %dw x %dh' % (grid.width, grid.height))
-    col_widths = []
-    for x in range(0, grid.width):
-        col_widths.append(max(len(bt.name) for bt in bg.block_types_for(col=x)))
-
-    print("    ", end=" ")
-    for x in range(0, grid.width):
-        print("{: ^{width}d}".format(x, width=col_widths[x]), end="   ")
-    print()
-
-    print("   /", end="-")
-    for x in range(0, grid.width):
-        print("-"*col_widths[x], end="-+-")
-    print()
-
-    for y in reversed(range(0, grid.height)):
-        print("{: 3d} |".format(y, width=col_widths[0]), end=" ")
-        for x, bt in enumerate(bg.block_types_for(row=y)):
-            assert x < len(col_widths), (x, bt)
-            print("{: ^{width}}".format(bt.name, width=col_widths[x]), end=" | ")
-        print()
-
-def print_nodes(rr_graph, lim=None):
-    '''Display source/sink edges on all XML nodes'''
-    ids = rr_graph.ids
-    print('Nodes: {}, edges {}'.format(len(ids._xml_nodes), len(ids._xml_edges)))
-    for nodei, node in enumerate(ids._xml_nodes):
-        print()
-        if lim and nodei >= lim:
-            print('...')
-            break
-        #print(nodei)
-        #ET.dump(node)
-        print('{} ({})'.format(ids.node_name(node), node.get("id")))
-        srcs = []
-        snks = []
-        for e in ids.edges_for_node(node):
-            src, snk = ids.nodes_for_edge(e)
-            if src == node:
-                srcs.append(e)
-            elif snk == node:
-                snks.append(e)
-            else:
-                print("!?@", ids.edge_name(e))
-
-        print("  Sources:")
-        for e in srcs:
-            print("   ", ids.edge_name(e))
-        if not srcs:
-            print("   ", None)
-
-        print("  Sink:")
-        for e in snks:
-            print("   ", ids.edge_name(e, flip=True))
-        if not snks:
-            print("   ", None)
-
-def print_graph(rr_graph, verbose=False, lim=10):
-    if verbose:
-        lim=0
-
-    print()
-    print_block_types(rr_graph)
-    print()
-    print_grid(rr_graph)
-    print()
-    print_nodes(rr_graph, lim=lim)
-    print()
-
 def main():
-    import os
-    if len(sys.argv) == 1 or not os.path.exists(sys.argv[-1]):
-        import doctest
-        print('Doctest begin')
-        doctest.testmod()
-        print('Doctest end')
-    else:
-        g = Graph(rr_graph_file=sys.argv[-1])
-        print_graph(g, verbose=True)
+    import doctest
+
+    print('Doctest begin')
+    doctest.testmod()
+    print('Doctest end')
 
 if __name__ == "__main__":
     main()
-
